requirements.txt
anthropic
pypdf
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
streamlit
python-dotenv
